requirements.txt
numpy<2
python-dotenv
python-osc
